requirements.txt
playwright>=1.40.0
pytest>=7.0.0
pytest-playwright>=0.4.0
allure-pytest>=2.13.0
